Fix crop images

+++ ServerlessCakesDetector.Core/Models/ImageRectangle.cs
@@ -8,9 +8,9 @@ namespace ServerlessCakesDetector.Core.Models
 {
 	public  class ImageRectangle
 	{
-		public int Left { get; set; }
-		public int Top { get; set; }
-		public int Width { get; set; }
-		public int Height { get; set; }
+		public double Left { get; set; }
+		public double Top { get; set; }
+		public double Width { get; set; }
+		public double Height { get; set; }
 	}
 }
